Messy Footer
This is the footer + most of the other changes we wanted. For the footer please make sure you check the welcome tab and the sources tab because it's wonky on those tabs. Buttons work though! They have no loading animation so it can be a little jarring if you click many of them in a row.

Co-Authored-By: terryguan97 <61436223+terryguan97@users.noreply.github.com>